refactor(chainconfig): remove hexagon icon
