+++ src/components/Icon/Chain/bsc.js
