+++ src/components/Icon/Chain/btc.js
