+++ src/components/Icon/Chain/cfx.js
