+++ src/constants/chainConfig.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import {IS_DEV} from './../utils'
-import BscHexagonIcon from './../components/Icon/Hexagon/bsc'
-import BtcHexagonIcon from './../components/Icon/Hexagon/btc'
-import EthHexagonIcon from './../components/Icon/Hexagon/eth'
-import CfxHexagonIcon from './../components/Icon/Hexagon/cfx'
+import BscIcon from './../components/Icon/Chain/bsc'
+import BtcIcon from './../components/Icon/Chain/btc'
+import EthIcon from './../components/Icon/Chain/eth'
+import CfxIcon from './../components/Icon/Chain/cfx'
 export const ChainNameEth = 'eth'
 export const ScanUrlEth = IS_DEV
   ? 'https://rinkeby.etherscan.io/'
@@ -20,29 +20,28 @@ export const ChainNameBtc = 'btc'
 export const ScanUrlBtc = IS_DEV
   ? 'https://blockstream.info/testnet'
   : 'https://blockstream.info'
-export function Hexagon({chain, width, height}) {
+export function Icon({chain, width, height}) {
   let comp = <></>
   switch (chain) {
     case ChainNameEth:
-      comp = <EthHexagonIcon width={width} height={height} />
+      comp = <EthIcon width={width} height={height} />
       break
     case ChainNameBsc:
-      comp = <BscHexagonIcon width={width} height={height} />
+      comp = <BscIcon width={width} height={height} />
       break
     case ChainNameCfx:
-      comp = <CfxHexagonIcon width={width} height={height} />
+      comp = <CfxIcon width={width} height={height} />
       break
     case ChainNameBtc:
-      comp = <BtcHexagonIcon width={width} height={height} />
+      comp = <BtcIcon width={width} height={height} />
       break
   }
   return comp
 }
 const config = {
   [ChainNameEth]: {
-    icon: '',
-    hexagonIcon(width, height) {
-      return <Hexagon width={width} height={height} chain={ChainNameEth} />
+    icon(width, height) {
+      return <Icon width={width} height={height} chain={ChainNameEth} />
     },
     fullName: 'Ethereum',
     shortName: 'eth',
@@ -65,9 +64,8 @@ const config = {
     commonTokens: ['ETH', 'USDT', 'eCFX'],
   },
   [ChainNameBsc]: {
-    icon: '',
-    hexagonIcon(width, height) {
-      return <Hexagon width={width} height={height} chain={ChainNameBsc} />
+    icon(width, height) {
+      return <Icon width={width} height={height} chain={ChainNameBsc} />
     },
     fullName: 'Binance Smart Contract',
     shortName: 'bsc',
@@ -90,9 +88,8 @@ const config = {
     commonTokens: ['BNB', 'bcUSDT', 'bCFX'],
   },
   [ChainNameCfx]: {
-    icon: '',
-    hexagonIcon(width, height) {
-      return <Hexagon width={width} height={height} chain={ChainNameCfx} />
+    icon(width, height) {
+      return <Icon width={width} height={height} chain={ChainNameCfx} />
     },
     fullName: 'Conflux',
     shortName: 'cfx',
@@ -115,9 +112,8 @@ const config = {
     commonTokens: ['CFX', 'cUSDT', 'cETH'],
   },
   [ChainNameBtc]: {
-    icon: '',
-    hexagonIcon(width, height) {
-      return <Hexagon width={width} height={height} chain={ChainNameBtc} />
+    icon(width, height) {
+      return <Icon width={width} height={height} chain={ChainNameBtc} />
     },
     fullName: 'Bitcoin',
     shortName: 'btc',
